% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling2.R
\name{add_clustered}
\alias{add_clustered}
\title{Add clustered standard errors and p-values}
\usage{
add_clustered(fit, cluster, type)
}
\arguments{
\item{fit}{linear model as returned by \code{lm()}}

\item{cluster}{vector which indicates the observations that should be clusterd.}

\item{type}{see \code{?clubSandwhich::vcovCR()}}
}
\value{
object of class \code{c("clustered", "lm")}
}
\description{
Will result in an object of class \code{clustered} (and \code{lm}).
}
