% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backtransform.R
\name{backtransform}
\alias{backtransform}
\title{Backtransforms the Response}
\usage{
backtransform(y)
}
\arguments{
\item{y}{log-transformed response}
}
\value{
numeric vector
}
\description{
Backtransforms the Response
}
