% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/default_data.R
\name{default_data}
\alias{default_data}
\title{Loads and Prepares the Data for Estimation}
\usage{
default_data()
}
\value{
Data frame
}
\description{
Loads and prepares the \code{\link{response_rates}} data for estimation. It selects the
input variables, computes the weights, rescales the responsee burden score and
log transforms the response rate.
}
