% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling2.R
\name{extract.clustered}
\alias{extract.clustered}
\title{#' @export
predict.clustered <- function(object, newdata, cluster = NULL, type = NULL) {
browser()
tmp <- object
class(tmp) <- "lm"
y_hat <- as.data.frame(predict.lm(object, newdata = as.data.frame(newdata), interval = "confidence"))
if (!is.null(cluster)) {
vcov_ <- vcov(tmp, cluster, type)
clustered_se_ <- clustered_se(newdata, object, vcov_)
critical_value_ <- critical_value(object)
lwr <- y_hat$fit - critical_value_ * clustered_se_
upr <- y_hat$fit + critical_value_ * clustered_se_
y_hat$lwr <- lwr
y_hat$upr <- upr
}
y_hat
}}
\usage{
extract.clustered(model)
}
\description{
#' @export
predict.clustered <- function(object, newdata, cluster = NULL, type = NULL) {
browser()
tmp <- object
class(tmp) <- "lm"
y_hat <- as.data.frame(predict.lm(object, newdata = as.data.frame(newdata), interval = "confidence"))
if (!is.null(cluster)) {
vcov_ <- vcov(tmp, cluster, type)
clustered_se_ <- clustered_se(newdata, object, vcov_)
critical_value_ <- critical_value(object)
lwr <- y_hat$fit - critical_value_ * clustered_se_
upr <- y_hat$fit + critical_value_ * clustered_se_
y_hat$lwr <- lwr
y_hat$upr <- upr
}
y_hat
}
}
