% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{data}
\name{response_burden_scheme}
\alias{response_burden_scheme}
\title{Response burden scheme}
\format{
Data frame with columns
\describe{
\item{Item}{Question item/type.}
\item{Points}{Response burden score.}
}
}
\usage{
response_burden_scheme
}
\description{
Elaborates on how to score different question items.
}
\keyword{datasets}
