% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/man.R
\docType{data}
\name{response_rates}
\alias{response_rates}
\title{Response rates data}
\format{
Data frame with columns
\describe{
\item{...}{TODO.}
}
}
\source{
Institute for Transport Planning and Systems (IVT), ETH Zurich,
\url{https://www.ivt.ethz.ch/en/}
}
\usage{
response_rates
}
\description{
Main data.frame to be updated on a rolling basis.
}
\keyword{datasets}
