% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/roxygen.R
\docType{package}
\name{responseRateAnalysis-package}
\alias{responseRateAnalysis}
\alias{responseRateAnalysis-package}
\title{responseRateAnalysis: A collective effort towards understanding response rates}
\description{
Organizes response burden score data collection and the analysis.
}
\author{
\strong{Maintainer}: Daniel Heimgartner \email{daniel.heimgartner@ivt.baug.ethz.ch} (\href{https://orcid.org/0000-0002-0643-8690}{ORCID}) [copyright holder]

}
