% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{summary.clustered}
\alias{summary.clustered}
\title{Summarizing Clustered Response Rate Analysis Objects}
\usage{
\method{summary}{clustered}(object, ...)
}
\arguments{
\item{object}{an object of class \verb{"clustered", "responseRateAnalysis"}.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
Called for side-effect.
}
\description{
Wrapper around \code{\link[texreg:screenreg]{texreg::screenreg}}.
}
