% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modeling.R
\name{extract,clustered-method}
\alias{extract,clustered-method}
\alias{extract.clustered}
\title{Extract Method for Clustered Response Rate Analysis Fits}
\usage{
\S4method{extract}{clustered}(model)
}
\arguments{
\item{model}{an object of class \verb{"clustered", "responseRateAnalysis"}.}
}
\value{
A \code{texreg-class} object representing the statistical model
}
\description{
This is the main method called when using functions from the \code{texreg-package}.
}
\details{
The \code{extract} method is called internally. Higher-level functions from the
\code{texreg-package} pass arguments via \code{...} to \code{extract}.
}
\seealso{
\code{texreg-package}, \code{\link[texreg:texreg]{texreg::texreg}}, \code{\link[texreg:screenreg]{texreg::screenreg}} and related functions.
}
